Fix compile.sh to work if UAC.rc is missing (eg launcher update rather than building an installer)
git-svn-id: http://svn.dmdirc.com/trunk@4437 00569f92-eb28-0410-84fd-f71c24880f43

+++ launcher/windows/compile.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 rm Vista.pas UAC.manifest UAC.rc
-cp ../../installer/windows/Vista.pas ../../installer/windows/UAC.manifest  ../../installer/windows/UAC.rc .
+if [ ! -e ../../installer/windows/UAC.rc ]; then
+	echo "1 24 \"UAC.manifest\"" > UAC.rc
+	cp ../../installer/windows/Vista.pas ../../installer/windows/UAC.manifest .
+else
+	cp ../../installer/windows/Vista.pas ../../installer/windows/UAC.manifest ../../installer/windows/UAC.rc .
+fi;
 
 rm -Rf ./*.exe
 fpc -Sd -Twin32 DMDirc.dpr
